fixed controllers

+++ src/main/java/com/example/courseregistrationsystem/controllers/CourseController.java
@@ -11,7 +11,7 @@ import java.util.*;
 @RequestMapping("/courses/")
 public class CourseController {
     @GetMapping("/{id}")
-    public Course getCourseById(@PathVariable("id") int id) {
+    public Course getCourseById(@PathVariable("id") Long id) {
         return null;
     }
     @GetMapping
@@ -20,12 +20,12 @@ public class CourseController {
     }
 
     @GetMapping("/department/{id}")
-    public List<Course> getCourseByDepartmentId(@PathVariable int id) {
+    public List<Course> getCourseByDepartmentId(@PathVariable Long id) {
         return null;
     }
 
     @GetMapping("/student/{id}")
-    public List<Course> getCourseByStudentId(@PathVariable int id) {
+    public List<Course> getCourseByStudentId(@PathVariable Long id) {
         return null;
     }
 
@@ -39,7 +39,7 @@ public class CourseController {
         return null;
     }
     @DeleteMapping("/{id}")
-    public void deleteCourse(@PathVariable("id") int id) {
+    public void deleteCourse(@PathVariable("id") Long id) {
         return;
     }
 }
